requirements.txt
Babel==2.6.0
colorama==0.4.1
dj-database-url==0.5.0
Django==2.1.7
django-appconf==1.0.3
django-axes==4.5.4
django-formtools==2.1
django-ipware==2.1.0
django-otp==0.5.2
django-phonenumber-field==1.3.0
django-two-factor-auth==1.8.0
django-widget-tweaks==1.4.3
gunicorn==19.7.1
numpy==1.14.2
phonenumberslite==8.10.8
Pillow==5.0.0
psycopg2==2.7.7
pytz==2018.3
qrcode==6.1
six==1.12.0
whitenoise==3.3.1
